feat: remove unused ButtonExpandNotifier and add CategoryButtonsWidget for enhanced category display
